OrderedDataConsumer

+++ se.redfield.knime.neo4jextension/se.redfield.knime.neo4jextension/src/main/java/se/redfield/knime/neo4j/async/AsyncRunner.java
@@ -17,7 +17,7 @@ public interface AsyncRunner<V> {
      * as simple throw it. It will correct catch and converted to script result with
      * exception and null result.
      */
-    void run(int number, V arg) throws Exception;
+    void run(long number, V arg) throws Exception;
     /**
      */
     default void workerStopped() {}
